backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
razorpay
httpx
python-dotenv
celery
redis
